sys: kern: arch: rv32: move saved_state structs to seperate module.
