+++ sys/kern/src/arch/riscv32/saved_state.rs
@@ -1,0 +1,117 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at https://mozilla.org/MPL/2.0/.
+
+use crate::task;
+use zerocopy::FromBytes;
+
+/// RISC-V volatile registers that must be saved across context switches.
+#[repr(C)]
+#[derive(Clone, Debug, Default, FromBytes)]
+pub struct SavedState {
+    // NOTE: the following fields must be kept contiguous!
+    ra: u32,
+    sp: u32,
+    gp: u32,
+    tp: u32,
+    t0: u32,
+    t1: u32,
+    t2: u32,
+    s0: u32,
+    s1: u32,
+    a0: u32,
+    a1: u32,
+    a2: u32,
+    a3: u32,
+    a4: u32,
+    a5: u32,
+    a6: u32,
+    a7: u32,
+    s2: u32,
+    s3: u32,
+    s4: u32,
+    s5: u32,
+    s6: u32,
+    s7: u32,
+    s8: u32,
+    s9: u32,
+    s10: u32,
+    s11: u32,
+    t3: u32,
+    t4: u32,
+    t5: u32,
+    t6: u32,
+    // Additional save value for task program counter
+    pc: u32,
+    // NOTE: the above fields must be kept contiguous!
+}
+
+impl SavedState {
+    pub fn sp(&self) -> u32 {
+        self.sp
+    }
+    pub fn pc(&self) -> u32 {
+        self.pc
+    }
+    pub fn set_sp(&mut self, val: u32) {
+        self.sp = val;
+    }
+    pub fn set_pc(&mut self, val: u32) {
+        self.pc = val;
+    }
+}
+
+/// Map the volatile registers to (architecture-independent) syscall argument
+/// and return slots.
+impl task::ArchState for SavedState {
+    fn stack_pointer(&self) -> u32 {
+        self.sp
+    }
+
+    /// Reads syscall argument register 0.
+    fn arg0(&self) -> u32 {
+        self.a0
+    }
+    fn arg1(&self) -> u32 {
+        self.a1
+    }
+    fn arg2(&self) -> u32 {
+        self.a2
+    }
+    fn arg3(&self) -> u32 {
+        self.a3
+    }
+    fn arg4(&self) -> u32 {
+        self.a4
+    }
+    fn arg5(&self) -> u32 {
+        self.a5
+    }
+    fn arg6(&self) -> u32 {
+        self.a6
+    }
+
+    fn syscall_descriptor(&self) -> u32 {
+        self.a7
+    }
+
+    /// Writes syscall return argument 0.
+    fn ret0(&mut self, x: u32) {
+        self.a0 = x
+    }
+    fn ret1(&mut self, x: u32) {
+        self.a1 = x
+    }
+    fn ret2(&mut self, x: u32) {
+        self.a2 = x
+    }
+    fn ret3(&mut self, x: u32) {
+        self.a3 = x
+    }
+    fn ret4(&mut self, x: u32) {
+        self.a4 = x
+    }
+    fn ret5(&mut self, x: u32) {
+        self.a5 = x
+    }
+}
